Change Swing Method to use model, Commit to update from master
