+++ ashraesizer.py
@@ -17,8 +17,7 @@
 
 """
 import numpy as np
-from cfg import rhoCp, TONS_TO_KBTUHR
-from HPWHComponents import mixVolume
+from cfg import rhoCp, TONS_TO_KBTUHR, mixVolume
 
 class ASHRAEsizer:
     """
