+++ cfg.py
@@ -30,3 +30,74 @@ Wapt75 = 1.75 # Real saftey factor used
 Wapt25 = 0.66
 TMSafetyFactor = 1
 compMinimumRunTime = 10./60.
+
+
+
+def mixVolume(vol, hotT, coldT, outT):
+    """
+    Adjusts the volume of water such that the hotT water and outT water have the
+    same amount of energy, meaning different volumes.
+
+    Parameters
+    ----------
+    vol : float
+        The reference volume to convert.
+    hotT : float
+        The hot water temperature used for mixing.
+    coldT : float
+        The cold water tempeature used for mixing.
+    outT : float
+        The out water temperature from mixing.
+
+    Returns
+    -------
+    float
+        Temperature adjusted volume.
+
+    """
+    fraction = (outT - coldT) / (hotT - coldT)
+
+    return vol * fraction
+
+
+def roundList(a_list, n=3):
+    """
+    Rounds elements in a python list
+
+    Parameters
+    ----------
+    a_list : float
+        list to round values of.
+    n : int
+        optional, default = 3. Number of digits to round elements to.
+
+    Returns
+    -------
+    list
+        rounded values.
+
+    """
+    return [round(num, n) for num in a_list]
+
+
+
+def HRLIST_to_MINLIST(a_list):
+    """
+    Repeats each element of a_list 60 times to go from hourly to minute.
+    Still may need other unit conversions to get data from per hour to per minute
+
+    Parameters
+    ----------
+    a_list : list
+        A list in of values per hour.
+
+    Returns
+    -------
+    out_list : list
+        A list in of values per minute created by repeating values per hour 60 times.
+
+    """
+    out_list = []
+    for num in a_list:
+        out_list += [num]*60
+    return out_list
