+++ tests/ref/ca_sizer.txt
@@ -1,8 +1,8 @@
 primarySystem:
 
-totalHWLoad, 2053.610
+totalHWLoad, 5077.923
 loadShapeNorm, [0.0152,0.0061,0.003 ,0.003 ,0.003 ,0.0091,0.076 ,0.152 ,0.1003,0.0821,0.0213,0.0243,0.0213,0.0243,0.0122,0.0061,0.0091,0.0365,0.0547,0.0729,0.0881,0.0912,0.0638,0.0243]
-nPeople, 78.582
+nPeople, 187.608
 incomingT_F, 50
 storageT_F, 150.000
 supplyT_F, 125
@@ -10,19 +10,20 @@ defrostFactor, 1
 percentUseable, 0.800
 compRuntime_hr, 16.000
 aquaFract, 0.400
-extraLoad_GPH, 14.705
+swingTank, <HPWHComponents.SwingTank object at 0x000001D34D6E5048>
+extraLoad_GPH, 35.945
 maxDayRun_hr, 16.000
 LS_on_off, [1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.,1.]
 loadShift, False
-PCap_kBTUhr, 94.233
-PVol_G_atStorageT, 542.647
-primaryCurve_vol, [829.91,815.45,800.24,784.23,767.35,749.52,730.66,710.68,689.49,666.96,642.96,617.34,589.95,560.58,529.01,494.99,458.21,418.34,374.96,336.72,299.07,270.47,238.81,203.58]
-primaryCurve_heatCap_kBTUhr, [ 62.82, 64.4 , 66.07, 67.82, 69.66, 71.61, 73.67, 75.86, 78.18, 80.64, 83.26, 86.06, 89.06, 92.27, 95.72, 99.44,103.47,107.83,112.57,117.75,123.43,129.68,136.61,144.31]
+PCap_kBTUhr, 232.617
+PVol_G_atStorageT, 1275.289
+primaryCurve_vol, [3137.14,2862.22,2573.1 ,2268.66,1947.63,1658.86,1623.88,1586.83,1547.51,1505.71,1461.2 ,1413.68,1362.86,1308.37,1249.81,1186.69,1118.46,1044.48, 963.99, 835.85, 763.56, 683.94, 597.74, 532.36, 459.15]
+primaryCurve_heatCap_kBTUhr, [155.08,158.98,163.09,167.41,171.97,176.78,181.88,187.27,192.99,199.08,205.56,212.47,219.87,227.8 ,236.33,245.51,255.45,266.21,277.93,290.73,304.76,320.21,337.31,356.35,377.66]
 
 temperatureMaintenanceSystem:
 schematic, swingtank
-nApt, 36
+nApt, 88
 Wapt, 100
-swingLoadToPrimary_W, 3600
-TMCap_kBTUhr, 21.496
-TMVol_G, 80
+swingLoadToPrimary_W, 8800
+TMCap_kBTUhr, 52.547
+TMVol_G, 120 - 300
